chore: rename to GTFluidUtils
